+++ DoctorWeb/DoctorWeb/Models/PatientHistory.cs
@@ -41,7 +41,7 @@ namespace DoctorWeb.Models
 
         public float? Weight { get; set; }
         public float? Height { get; set; }
-        public int? T { get; set; }
+        public float? T { get; set; }
         public string PR { get; set; }
         public string BP { get; set; }
         public string SPO2 { get; set; }
